+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ## Fixes
 
-* Oid underlying type changed to uint32, previously it was incorrectly int32
+* Oid underlying type changed to uint32, previously it was incorrectly int32 (Manni Wood)
+
+## Features
+
+* Add xid type support (Manni Wood)
+* Add cid type support (Manni Wood)
+* Add tid type support (Manni Wood)
+* Add "char" type support (Manni Wood)
+* Add NullOid type (Manni Wood)
+* Add json/jsonb binary support to allow use with CopyTo
 
 # 2.9.0 (August 26, 2016)